# docs/services/gateway/api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Power Knowledge Platform Gateway API
  version: 0.1.0
  license:
    name: Unspecified
    url: https://github.com/Sakayori-Iroha-168/Software_Teamwork
  description: |
    Public caller-facing API contract for the gateway service.
    All stable public and service-to-service HTTP APIs must use RESTful resource-oriented paths.
    This revision treats gateway health checks, user/session,
    knowledge-owned knowledge base, document upload, processing, content, chunk, and retrieval routes,
    document-owned report generation and report settings routes, qa-owned intelligent question answering
    routes, admin-facing runtime model/parser configuration routes, admin user
    management, personal profile, required password-change routes, and gateway-owned
    app version freshness routes as active contracts. Both admin-overview and
    admin-metrics are active gateway contracts for the administration dashboard.
...
servers:
- url: /
  description: Gateway root
security:
- bearerAuth: []
...
paths:
...
  /api/v1/qa-config-versions:
    post:
      tags:
      - qa-settings
      summary: Create QA config version
      operationId: createQAConfigVersion
      x-owner-service: qa
      x-required-permissions:
        - qa:settings:write
      description: >
        Creates a versioned QA runtime configuration so historical answers remain
        traceable to the retrieval, Agent settings, and system prompt used at
        generation time. Requires qa:settings:write permission. The created
        version (including systemPrompt) is returned on success. The published
        system prompt becomes active for the next question; in-flight answer
        generations continue using their previously acquired snapshot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQAConfigVersionRequest'
      responses:
        '201':
          description: QA config version created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QAConfigVersionResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token issued by auth and cached by gateway through a token hash; callers must not treat it as a JWT.
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          enum:
          - validation_error
          - unauthorized
          - forbidden
          - not_found
          - conflict
          - rate_limited
          - dependency_error
          - internal_error
          - not_implemented
          - unsupported_intent
          - unsupported_mode
        message:
          type: string
        requestId:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...
    QARetrievalOptions:
      type: object
      properties:
        topK:
          type: integer
          minimum: 1
          maximum: 100
        scoreThreshold:
          type: number
          minimum: 0
        similarityThreshold:
          type: number
          minimum: 0
          deprecated: true
          description: Backward-compatible alias for `scoreThreshold`.
        enableRerank:
          type: boolean
        useRerank:
          type: boolean
          deprecated: true
          description: Backward-compatible alias for `enableRerank`.
        rerankThreshold:
          type: number
          minimum: 0
        rerankTopN:
          type: integer
          minimum: 1
        tagFilters:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    QAConfigKnowledgeBase:
      type: object
      required:
      - id
      properties:
        id:
          type: string
        type:
          type: string
        displayName:
          type: string
        sortOrder:
          type: integer
          minimum: 0
    QAConfigVersion:
      type: object
      required:
      - id
      - versionNo
      - systemPrompt
      - isActive
      - createdAt
      properties:
        id:
          type: string
        versionNo:
          type: integer
          minimum: 1
        defaultKnowledgeBaseIds:
          type: array
          items:
            type: string
          description: >
            Optional default QA retrieval allowlist. An empty list means QA does
            not narrow retrieval and uses the project-wide knowledge pool.
        knowledgeBases:
          type: array
          items:
            $ref: '#/components/schemas/QAConfigKnowledgeBase'
        retrieval:
          $ref: '#/components/schemas/QARetrievalOptions'
        maxIterations:
          type: integer
          minimum: 1
        toolTimeoutSeconds:
          type: integer
          minimum: 1
        modelTimeoutSeconds:
          type: integer
          minimum: 1
        overallTimeoutSeconds:
          type: integer
          minimum: 1
        enabledToolNames:
          type: array
          items:
            type: string
        llm:
          $ref: '#/components/schemas/QALLMConfigVersion'
        agent:
          $ref: '#/components/schemas/QAAgentConfig'
        systemPrompt:
          type: string
          description: >
            Global Agent system prompt. Returned only in admin QA config
            endpoints (requires qa:settings:read). Must not appear in SSE
            events, QA chat responses, logs, metrics, tool call summaries,
            or error messages.
            Server validates 1–20000 bytes (octet_length, not character count);
            multi-byte UTF-8 characters count toward the byte limit.
        isActive:
          type: boolean
        createdBy:
          type: string
          description: User id that created this QA config version.
        createdAt:
          type: string
          format: date-time
    QAAgentConfig:
      type: object
      properties:
        maxIterations:
          type: integer
          minimum: 1
          default: 5
        toolTimeoutSeconds:
          type: integer
          minimum: 1
          default: 10
        modelTimeoutSeconds:
          type: integer
          minimum: 1
          default: 60
        overallTimeoutSeconds:
          type: integer
          minimum: 1
          default: 120
        enabledToolNames:
          type: array
          items:
            type: string
    CreateQAConfigVersionRequest:
      type: object
      properties:
        defaultKnowledgeBaseIds:
          type: array
          items:
            type: string
          description: >
            Optional default QA retrieval allowlist. An empty list means QA does
            not narrow retrieval and uses the project-wide knowledge pool.
        knowledgeBases:
          type: array
          items:
            $ref: '#/components/schemas/QAConfigKnowledgeBase'
        retrieval:
          $ref: '#/components/schemas/QARetrievalOptions'
        maxIterations:
          type: integer
          minimum: 1
        toolTimeoutSeconds:
          type: integer
          minimum: 1
        modelTimeoutSeconds:
          type: integer
          minimum: 1
        overallTimeoutSeconds:
          type: integer
          minimum: 1
        enabledToolNames:
          type: array
          items:
            type: string
        agent:
          $ref: '#/components/schemas/QAAgentConfig'
        systemPrompt:
          type: string
          description: >
            Global Agent system prompt. Optional on create; if omitted, the
            current active version's prompt is inherited so that only changing
            retrieval/agent settings does not clear the existing prompt.
            Server validates 1–20000 bytes (octet_length, not character count)
            after trimming; multi-byte UTF-8 characters count toward the byte
            limit. Clients should apply the same byte-length check before
            submitting to avoid 400 validation errors.
        activate:
          type: boolean
          default: true
    QAConfigVersionResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/QAConfigVersion'
        requestId:
          type: string
    QALLMConfigVersion:
      type: object
      required:
      - id
      - versionNo
      - provider
      - profileId
      - modelName
      - isActive
      - createdAt
      properties:
        id:
          type: string
        versionNo:
          type: integer
          minimum: 1
        provider:
          type: string
          enum:
          - ai-gateway
        profileId:
          type: string
          description: AI Gateway chat model profile id.
        modelName:
          type: string
        timeoutSeconds:
          type: integer
          minimum: 1
        temperature:
          type: number
        maxTokens:
          type: integer
          minimum: 1
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...